Remove useState from server rendered page
Remove unused application page
Remove unused loading spinners
Remove apply button from finance view
Add space to subscribe form
Remove parameters to testimonial component
Remove career from queries
Match finance and skill queries to job
