+++ lib/queries.js
@@ -4,8 +4,7 @@ export const findJobById = (id) => `{
 
 export const findOpportunities = (count = "5") => `{
   "job": *[_type == "job" && dateTime(_updatedAt) > dateTime(now()) - 60*60*24*7 && !(_id in path('drafts.**'))]{_id,_createdAt,position,description,location,qualifications,closingDate,enableApply,link,responsibilities,role,slug,title,companyRef->{company,bio,logo,website}} | order(_createdAt desc)[0..${count}],
-  "finance": *[_type == "finance" && !(_id in path('drafts.**'))]{_id,_createdAt,title,description,companyRef->{company,logo}} | order(_createdAt desc)[0..${count}],
-  "career": *[_type == "career" && !(_id in path('drafts.**'))]{_id,_createdAt,companyRef->{company,logo}} | order(_createdAt desc)[0..${count}],
-  "skill": *[_type == "skill" && !(_id in path('drafts.**'))]{_id,_createdAt,title,description,companyRef->{company,logo}} | order(_createdAt desc)[0..${count}],
+  "finance": *[_type == "finance" && dateTime(_updatedAt) > dateTime(now()) - 60*60*24*7 && !(_id in path('drafts.**'))]{_id,_createdAt,position,description,location,qualifications,closingDate,enableApply,link,responsibilities,role,slug,title,companyRef->{company,bio,logo,website}} | order(_createdAt desc)[0..${count}],
+  "skill": *[_type == "skill" && dateTime(_updatedAt) > dateTime(now()) - 60*60*24*7 && !(_id in path('drafts.**'))]{_id,_createdAt,position,description,location,qualifications,closingDate,enableApply,link,responsibilities,role,slug,title,companyRef->{company,bio,logo,website}} | order(_createdAt desc)[0..${count}],
   "banner": *[_type == "banner"][0..${count}]
 }`;
